configuração inicial do projeto

+++ src/main/java/br/dev/guilhermecordeiro/consulta_cpfcnpj/integration/credlink/dtos/Trabalho.java
@@ -2,9 +2,11 @@ package br.dev.guilhermecordeiro.consulta_cpfcnpj.integration.credlink.dtos;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Trabalho {
     private String dataadmissao;
     private String pis;
-    private int rendaestimada;
+    private BigDecimal rendaestimada;
 }
